correct several places for color distribution
move the ControlColor related func to main
correct the handler of apply and undo on NORMAL txn
